+++ tailwind.config.js
@@ -3,8 +3,10 @@ const plugin = require("tailwindcss/plugin");
 module.exports = {
   purge: {
     content: [
+      "./src/assets/**/*.{js,ts,jsx,tsx}",
       "./src/components/**/*.{js,ts,jsx,tsx}",
       "./src/stories/**/*.{js,ts,jsx,tsx}",
+      "./src/assets/**/*.{js,ts,jsx,tsx}",
       "./config-plugins/**/*.{js,ts,jsx,tsx}",
     ],
   },
@@ -39,6 +41,9 @@ module.exports = {
         browserh9: ["17px", "26px"],
         browserbody: ["20px", "33px"],
         browserbodybold: ["20px", "33px"],
+
+        btnsupertask: ["22px", "32px"],
+        btnother: ["20px", "23px"],
       },
       fontWeight: {
         regular: 400,
@@ -241,9 +246,12 @@ module.exports = {
   },
   variants: {
     extend: {
-      backgroundColor: ["active"],
+      ringOffsetWidth: ["active", "hover", "focus"],
+      ringColor: ["active", "hover", "focus"],
+      backgroundColor: ["active", "hover", "focus"],
       textColor: ["visited"],
       margin: ["checked"],
+      textDecoration: ["active", "hover", "visited"],
     },
   },
   plugins: [
@@ -264,5 +272,7 @@ module.exports = {
     }),
     require("./config-plugins/typography-plugin"),
     require("./config-plugins/divider-plugin"),
+    require("./config-plugins/button-plugin"),
+    require("./config-plugins/errorpage-plugin"),
   ],
 };