+++ CHANGELOG.md
@@ -16,6 +16,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - Heading component
 - Text Link Component
 - Documentation and styling for Dividers, Colors, Typography
+- Button test cases/stories/html (react.js)
+- Image Component
+- Link Component
 
 ### Changed